+++ docs/changelog.rst
@@ -144,6 +144,8 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - Wayland: Fix key repeat being stopped by the release of an unrelated key
   (:iss:`2191`)
 
+- Wayland: Add support for the text input protocol (:iss:`3410`)
+
 - Add an option, :opt:`detect_urls` to control whether kitty will detect URLs
   when the mouse moves over them (:pull:`3118`)